+++ packages/react/src/RateButton/templates/Star/index.tsx
@@ -1,0 +1,12 @@
+/** @jsxRuntime classic */
+/* @jsx jsx */
+
+import { jsx } from "theme-ui";
+import { RateButtonTemplateComponentProps } from "../..";
+import { StarIcon } from "./icons/StarIcon";
+import { FC } from "react";
+import { Base } from "../Base";
+
+export const Star: FC<RateButtonTemplateComponentProps> = ({ ...props }) => {
+	return <Base {...props} Icon={StarIcon} />;
+};